+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.15.3] - 2026-05-19
+
+### Changed
+- update find_outliers with auto quantile type selection (type 2 for small samples, type 7 for large)
+- Updated files: R/find_outliers.R
+
+
 ## [1.15.2] - 2026-05-18
 
 ### Changed